backend/requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
websocket
python-multipart

# Claude Agent SDK (install from GitHub: pip install git+https://github.com/anthropics/claude-agent-sdk-python.git)
# claude-agent-sdk>=0.1.9

# Docker
docker

claude_agent_sdk

# GitHub
PyGithub
gitpython

# Database
sqlalchemy
aiosqlite

# Configuration
python-dotenv
pydantic-settings
pyyaml

# Utilities
aiofiles
httpx

